factor generator-extensions into separate libraries
