+++ libs/shared/tsconfig.libs.json
@@ -12,7 +12,9 @@
       "@bitwarden/components": ["../components/src"],
       "@bitwarden/generator-components": ["../tools/generator/components/src"],
       "@bitwarden/generator-core": ["../tools/generator/core/src"],
-      "@bitwarden/generator-extensions": ["../tools/generator/extensions/src"],
+      "@bitwarden/generator-history": ["../tools/generator/extensions/history/src"],
+      "@bitwarden/generator-legacy": ["../tools/generator/extensions/legacy/src"],
+      "@bitwarden/generator-navigation": ["../tools/generator/extensions/navigation/src"],
       "@bitwarden/vault-export-core": ["../tools/export/vault-export/vault-export-core/src"],
       "@bitwarden/vault-export-ui": ["../tools/export/vault-export/vault-export-ui/src"],
       "@bitwarden/importer/core": ["../importer/src"],
