+++ libs/tools/generator/components/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@bitwarden/generator-components",
   "version": "0.0.0",
-  "description": "Angular components for the Bitwarden generators",
+  "description": "Angular components for the Bitwarden credential generators",
   "keywords": [
     "bitwarden"
   ],
