+++ libs/tools/generator/core/src/abstractions/index.ts
@@ -1,5 +1,3 @@
-export { GeneratorHistoryService } from "../../../extensions/src/history/generator-history.abstraction";
-export { GeneratorNavigationService } from "../../../extensions/src/navigation/generator-navigation.service.abstraction";
 export { GeneratorService } from "./generator.service.abstraction";
 export { GeneratorStrategy } from "./generator-strategy.abstraction";
 export { PolicyEvaluator } from "./policy-evaluator.abstraction";
