+++ libs/tools/generator/extensions/history/jest.config.js
@@ -1,13 +1,13 @@
 const { pathsToModuleNameMapper } = require("ts-jest");
 
-const { compilerOptions } = require("../../../shared/tsconfig.libs");
+const { compilerOptions } = require("../../../../shared/tsconfig.libs");
 
 /** @type {import('jest').Config} */
 module.exports = {
   testMatch: ["**/+(*.)+(spec).+(ts)"],
   preset: "ts-jest",
-  testEnvironment: "../../../shared/test.environment.ts",
+  testEnvironment: "../../../../shared/test.environment.ts",
   moduleNameMapper: pathsToModuleNameMapper(compilerOptions?.paths || {}, {
-    prefix: "<rootDir>/../../",
+    prefix: "<rootDir>/../../../",
   }),
 };
