+++ libs/tools/generator/extensions/history/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@bitwarden/generator-core",
+  "name": "@bitwarden/generator-history",
   "version": "0.0.0",
-  "description": "Common Bitwarden credential generation logic",
+  "description": "Bitwarden credential generator history service",
   "keywords": [
     "bitwarden"
   ],
@@ -18,6 +18,7 @@
     "build:watch": "npm run clean && tsc -watch"
   },
   "dependencies": {
-    "@bitwarden/common": "file:../../../common"
+    "@bitwarden/common": "file:../../../common",
+    "@bitwarden/generator-core": "file:../core"
   }
 }
