+++ libs/tools/generator/extensions/history/src/generated-credential.ts
