+++ libs/tools/generator/extensions/history/src/generated-password-history.ts
