+++ libs/tools/generator/extensions/history/src/generator-history.abstraction.ts
