+++ libs/tools/generator/extensions/history/src/key-definition.spec.ts
