+++ libs/tools/generator/extensions/history/src/key-definitions.ts
