+++ libs/tools/generator/extensions/history/src/legacy-password-history-decryptor.ts
