+++ libs/tools/generator/extensions/history/src/options.ts
