+++ libs/tools/generator/extensions/legacy/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@bitwarden/generator-extensions",
+  "name": "@bitwarden/generator-legacy",
   "version": "0.0.0",
-  "description": "TODO",
+  "description": "Deprecated Bitwarden credential generation services",
   "keywords": [
     "bitwarden"
   ],
