+++ libs/tools/generator/extensions/legacy/src/index.ts
@@ -1,0 +1,6 @@
+export * from "./password-generation.service.abstraction";
+export * from "./create-legacy-password-generation-service";
+export * from "./password-generator-options";
+export * from "./username-generation.service.abstraction";
+export * from "./create-legacy-username-generation-service";
+export * from "./username-generation-options";
