+++ libs/tools/generator/extensions/legacy/src/password-generator-options.ts
@@ -1,6 +1,5 @@
 import { PassphraseGenerationOptions, PasswordGenerationOptions } from "@bitwarden/generator-core";
-
-import { GeneratorNavigation } from "../navigation";
+import { GeneratorNavigation } from "@bitwarden/generator-navigation";
 
 /** Request format for credential generation.
  *  This type includes all properties suitable for reactive data binding.
