+++ libs/tools/generator/extensions/legacy/src/username-generation-options.ts
