+++ libs/tools/generator/extensions/legacy/src/username-generation.service.abstraction.ts
