+++ libs/tools/generator/extensions/legacy/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../../../shared/tsconfig.libs",
+  "include": ["src"],
+  "exclude": ["node_modules", "dist"]
+}
