+++ libs/tools/generator/extensions/legacy/tsconfig.spec.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./tsconfig.json"
+}
