+++ libs/tools/generator/extensions/navigation/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@bitwarden/generator-core",
+  "name": "@bitwarden/generator-navigation",
   "version": "0.0.0",
-  "description": "Common Bitwarden credential generation logic",
+  "description": "Bitwarden credential generator selection settings service",
   "keywords": [
     "bitwarden"
   ],
@@ -18,6 +18,7 @@
     "build:watch": "npm run clean && tsc -watch"
   },
   "dependencies": {
-    "@bitwarden/common": "file:../../../common"
+    "@bitwarden/common": "file:../../../common",
+    "@bitwarden/generator-core": "file:../core"
   }
 }
