+++ libs/tools/generator/extensions/navigation/src/default-generator-navigation.ts
