+++ libs/tools/generator/extensions/navigation/src/generator-navigation-evaluator.spec.ts
