+++ libs/tools/generator/extensions/navigation/src/generator-navigation-evaluator.ts
