+++ libs/tools/generator/extensions/navigation/src/generator-navigation-policy.spec.ts
