+++ libs/tools/generator/extensions/navigation/src/generator-navigation.service.abstraction.ts
