+++ libs/tools/generator/extensions/navigation/src/generator-navigation.ts
