+++ libs/tools/generator/extensions/navigation/src/index.ts
