+++ libs/tools/generator/extensions/navigation/src/key-definition.spec.ts
