+++ libs/tools/generator/extensions/navigation/src/key-definitions.ts
