+++ libs/tools/generator/extensions/navigation/tsconfig.spec.json
@@ -1,0 +1,3 @@
+{
+  "extends": "./tsconfig.json"
+}
