--- libs/tools/generator/extensions/src/index.ts
@@ -1,9 +1,0 @@
-export * as history from "./history";
-export * as legacyPassword from "./legacy-password";
-export * as legacyUsername from "./legacy-username";
-export * as navigation from "./navigation";
-
-export { GeneratorHistoryService } from "./history";
-export { GeneratorNavigationService } from "./navigation";
-export { PasswordGenerationServiceAbstraction } from "./legacy-password";
-export { UsernameGenerationServiceAbstraction } from "./legacy-username";
