--- libs/tools/generator/extensions/src/legacy-password/index.ts
@@ -1,3 +1,0 @@
-export * from "./password-generation.service.abstraction";
-export * from "./create-generation-service";
-export * from "./password-generator-options";
