--- libs/tools/generator/extensions/src/legacy-username/index.ts
@@ -1,3 +1,0 @@
-export * from "./username-generation.service.abstraction";
-export * from "./create-generation-service";
-export * from "./username-generation-options";
